% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoids.R
\name{get_panoids}
\alias{get_panoids}
\title{get_panoids}
\usage{
get_panoids(path.root = NULL, query.addresses = NULL, n.echo = 10)
}
\arguments{
\item{path.root}{location of current dropbox root Example: ~/Dropbox/pkg.data/}

\item{query.addresses}{character vector of locations to find nearest panoids Example: 11804 Ridge Pkwy, Broomfield, CO 80021}

\item{n.echo}{Number of queries before backup and display}
}
\description{
Retrieve list of panoids and locations
}
\keyword{google}
\keyword{panoids}
\keyword{streetview}
