% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoids.R
\name{plot_panoids}
\alias{plot_panoids}
\title{plot_panoids}
\usage{
plot_panoids(path.root, l.extent = NULL)
}
\arguments{
\item{path.root}{location of current dropbox root Example: ~/Dropbox/pkg.data/}

\item{l.extent}{list l.extent$ lng.min, lng.max, lat.min,lat.max}
}
\description{
Plot panoids by snapshot date
}
\keyword{google}
\keyword{panoids}
\keyword{plot}
\keyword{streetview}
